backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
livekit==0.11.0
livekit-api==0.5.0
pydantic==2.5.0
